* commit 'b8427404e4d959d9b1f5ba5f454dd2b3c667f202':
  MIES_DAC-Hardware.ipf: Fix function signatures used without ITC XOP2
  ManualInstallation.md: Reorganize some topics
  HW_ITC_CloseAllDevices: Stop all running devices
  Hardware Tests: Set interactive mode to false in manual initialization routine
  Testing with hardware: Prefer HW wrapper function
  clean_mies_installation*: Add option to skip XOP installation
  CompilationTesting: Remove compilation testing for the separate modules
  ManualInstallation.md: Adapt installation instructions
  Remove module support
  Blowout: Add proper module name

Original-CommitID: ece4ced3c6159b6288c27a4caf673c56fba1261c

+++ tools/compilation-testing/clean_mies_installation.sh
@@ -6,6 +6,13 @@
 
 set -e
 
+if [ "$#" -gt 0 -a "$1" = "skipHardwareXOPs" ]
+then
+  installHWXOPs=0
+else
+  installHWXOPs=1
+fi
+
 git --version > /dev/null
 if [ $? -ne 0 ]
 then
@@ -36,13 +43,21 @@ user_proc="$IGOR_USER_FILES/User Procedures"
 xops="$IGOR_USER_FILES/Igor Extensions (64-bit)"
 
 mkdir -p "$user_proc"
-mkdir -p "$xops"
 
 rm -rf "$top_level"/Packages/doc/html
 cp -r  "$top_level"/Packages/*  "$user_proc"
-cp -r  "$top_level"/XOPs-IP7-64bit/*  "$xops"
-cp -r  "$top_level"/XOP-tango-IP7-64bit/* "$xops"
-rm -f  "$xops"/NIDAQmx64.*
+
+mkdir -p "$xops"
+
+if [ "$installHWXOPs" = "1" ]
+then
+  cp -r  "$top_level"/XOPs-IP7-64bit/*  "$xops"
+  cp -r  "$top_level"/XOP-tango-IP7-64bit/* "$xops"
+  rm -f  "$xops"/NIDAQmx64.*
+else
+  cp -r  "$top_level"/XOPs-IP7-64bit/HDF5*  "$xops"
+fi
+
 echo "Release: FAKE MIES VERSION" > "$IGOR_USER_FILES"/version.txt
 
 exit 0
